+++ src/main/java/com/bereznev/model/Salary.java
@@ -9,11 +9,13 @@ import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
+
 @Data
 @NoArgsConstructor
 @AllArgsConstructor
 public class Salary {
-    private int from;
-    private int to;
+    private BigDecimal from;
+    private BigDecimal to;
     private String currency;
 }
